template update

+++ Template.Application/Contracts/Persistence/IGenericRepository.cs
@@ -4,7 +4,7 @@ namespace Template.Application.Contracts.Persistence
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> GetAsync(int id);
+        Task<T> GetAsync(long id);
         Task<IReadOnlyList<T>> GetAllAsync();
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
